services/accept-order-saga-orchestrator/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.6
requests==2.28.2
python-dotenv==1.0.0
confluent-kafka==2.1.1
Werkzeug==2.2.3
kafka-python==2.0.2
